+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Polyverse Boost ReST API (Backend)
 
 ### Bug Fixes
 - Always return Resource Status (with last_updated time) for Resources that exist /api/user_project/{org}/{project}/data/{resource}/status (GET) - was returning no resource if no timestamp was stored
+- Fix JSON/String decoding from buffer for Generator POST/PUT/PATCH and process (was failing to parse)
 
 ## Version 0.9.7: January 30th, 2024